requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
python-multipart==0.0.12
groq>=0.9.0
google-generativeai>=0.8.0
supabase>=2.0.0
PyJWT>=2.8.0
faster-whisper>=1.0.0
